+++ Projects/Concert/ConcertDataTests/Sources/ConcertMapperTests.swift
@@ -176,204 +176,4 @@ struct ConcertMapperTests {
         #expect(result.debutYear == "2020")
         #expect(result.keywords.count == 2)
     }
-
-    // MARK: - Comment Tests
-
-    @Test("댓글 목록 매핑")
-    func testFetchConcertCommentListMapping() {
-        // Given
-        let response = DTO.Response.FetchConcertCommentList(
-            data: [
-                DTO.Response.FetchConcertCommentList.Comment(
-                    id: 1,
-                    userID: 100,
-                    nickname: "유저1",
-                    concertID: 1,
-                    content: "기대돼요!",
-                    createdAt: "2025-01-01T12:00:00Z"
-                )
-            ],
-            cursor: DTO.Response.FetchConcertCommentList.Cursor(createdAt: "2025-01-01T12:00:00Z", id: 1),
-            totalCount: 1
-        )
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result.comments.count == 1)
-        #expect(result.comments[0].nickname == "유저1")
-        #expect(result.totalCount == 1)
-        #expect(result.cursor != nil)
-    }
-
-    @Test("댓글 생성 매핑")
-    func testCreateConcertCommentMapping() {
-        // Given
-        let response = DTO.Response.CreateConcertComment(
-            id: 1,
-            userID: 100,
-            nickname: "유저1",
-            concertID: 1,
-            content: "댓글 내용",
-            createdAt: "2025-01-01T12:00:00Z"
-        )
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result.id == 1)
-        #expect(result.userID == 100)
-        #expect(result.nickname == "유저1")
-        #expect(result.content == "댓글 내용")
-    }
-
-    // MARK: - Setlist Tests
-
-    @Test("셋리스트 목록 매핑")
-    func testFetchConcertSetlistListMapping() {
-        // Given
-        let response: DTO.Response.FetchConcertSetlistList = [
-            DTO.Response.ConcertSetlist(
-                id: 1,
-                title: "2025 콘서트 Day1",
-                imageURL: "https://example.com/setlist.jpg",
-                type: "ONGOING",
-                startDate: "2025-01-01T18:00:00Z",
-                endDate: "2025-01-01T21:00:00Z",
-                status: "expected",
-                venue: "올림픽공원",
-                artist: "아티스트"
-            )
-        ]
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result.count == 1)
-        #expect(result[0].id == 1)
-        #expect(result[0].title == "2025 콘서트 Day1")
-        #expect(result[0].type == .ongoing)
-        #expect(result[0].status == .expected)
-    }
-
-    @Test("셋리스트 status가 nil일 때 none으로 매핑")
-    func testFetchConcertSetlistListMappingWithNilStatus() {
-        // Given
-        let response: DTO.Response.FetchConcertSetlistList = [
-            DTO.Response.ConcertSetlist(
-                id: 1,
-                title: "2025 콘서트",
-                imageURL: nil,
-                type: "ONGOING",
-                startDate: "2025-01-01T18:00:00Z",
-                endDate: "2025-01-01T21:00:00Z",
-                status: nil,
-                venue: "올림픽공원",
-                artist: "아티스트"
-            )
-        ]
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result.count == 1)
-        #expect(result[0].status == .none)
-    }
-
-    @Test("잘못된 타입으로 셋리스트 매핑 실패")
-    func testFetchConcertSetlistListMappingWithInvalidType() {
-        // Given
-        let response: DTO.Response.FetchConcertSetlistList = [
-            DTO.Response.ConcertSetlist(
-                id: 1,
-                title: "2025 콘서트",
-                imageURL: nil,
-                type: "INVALID_TYPE",
-                startDate: "2025-01-01T18:00:00Z",
-                endDate: "2025-01-01T21:00:00Z",
-                status: "expected",
-                venue: "올림픽공원",
-                artist: "아티스트"
-            )
-        ]
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result.isEmpty)
-    }
-
-    // MARK: - Setlist Detail Tests
-
-    @Test("셋리스트 상세 매핑")
-    func testFetchConcertSetlistDetailMapping() {
-        // Given
-        let response = DTO.Response.FetchConcertSetlist(
-            id: 1,
-            title: "2025 콘서트 Day1",
-            imageURL: "https://example.com/setlist.jpg",
-            type: "ONGOING",
-            startDate: "2025-01-01T18:00:00Z",
-            endDate: "2025-01-01T21:00:00Z",
-            status: "recent",
-            venue: "올림픽공원",
-            artist: "아티스트"
-        )
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result != nil)
-        #expect(result?.id == 1)
-        #expect(result?.type == .ongoing)
-        #expect(result?.status == .recent)
-    }
-
-    @Test("잘못된 날짜로 셋리스트 상세 매핑 실패")
-    func testFetchConcertSetlistDetailMappingWithInvalidDate() {
-        // Given
-        let response = DTO.Response.FetchConcertSetlist(
-            id: 1,
-            title: "2025 콘서트",
-            imageURL: nil,
-            type: "ONGOING",
-            startDate: "invalid-date",
-            endDate: "2025-01-01T21:00:00Z",
-            status: "expected",
-            venue: "올림픽공원",
-            artist: "아티스트"
-        )
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result == nil)
-    }
-
-    // MARK: - Setlist Song Tests
-
-    @Test("셋리스트 곡 목록 매핑")
-    func testFetchSetlistSongListMapping() {
-        // Given
-        let response: DTO.Response.FetchSetlistSongList = [
-            DTO.Response.SetlistSong(id: 1, title: "곡1", artist: "아티스트", orderIndex: 1),
-            DTO.Response.SetlistSong(id: 2, title: "곡2", artist: "아티스트", orderIndex: 2)
-        ]
-
-        // When
-        let result = mapper.toDomain(from: response)
-
-        // Then
-        #expect(result.count == 2)
-        #expect(result[0].title == "곡1")
-        #expect(result[0].orderIndex == 1)
-        #expect(result[1].orderIndex == 2)
-    }
 }
